CourseAssistant/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
itsdangerous==2.1.2
bcrypt==4.1.3
sentence-transformers==2.7.0
faiss-cpu==1.9.0.post1; platform_system == "Windows"
faiss-cpu==1.8.0; platform_system != "Windows"
PyMuPDF==1.24.3
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pytesseract==0.3.10
Pillow==10.3.0
chardet==5.2.0
matplotlib==3.8.4
numpy==1.26.4
httpx==0.27.0
aiosqlite==0.20.0
python-dotenv==1.0.1
sse-starlette==1.6.5
